# back/central-reserve/internal/infra/primary/http2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/businesses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene la lista completa de negocios disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Obtener todos los negocios",
                "responses": {
                    "200": {
                        "description": "Negocios obtenidos exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessListSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo negocio con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Crear negocio",
                "parameters": [
                    {
                        "description": "Datos del negocio",
                        "name": "business",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BusinessRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Negocio creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Código o dominio ya existe",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    }
                }
            }
        },
        "/businesses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un negocio específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Obtener negocio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio obtenido exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un negocio existente con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Actualizar negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del negocio",
                        "name": "business",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BusinessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Código o dominio ya existe",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un negocio existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Eliminar negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio eliminado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_infra_primary_http2_handlers_businesshandler_response.BusinessTypeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.BusinessRequest": {
            "type": "object",
            "required": [
                "business_type_id",
                "code",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "business_type_id": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "custom_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable_delivery": {
                    "description": "Configuración de funcionalidades",
                    "type": "boolean"
                },
                "enable_pickup": {
                    "type": "boolean"
                },
                "enable_reservations": {
                    "type": "boolean"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "description": "Configuración de marca blanca",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary_color": {
                    "type": "string"
                },
                "secondary_color": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "response.BusinessErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.BusinessListResponse": {
            "type": "object",
            "properties": {
                "businesses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.BusinessResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.BusinessListSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.BusinessListResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.BusinessResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "business_type": {
                    "$ref": "#/definitions/internal_infra_primary_http2_handlers_businesshandler_response.BusinessTypeResponse"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "custom_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable_delivery": {
                    "description": "Configuración de funcionalidades",
                    "type": "boolean"
                },
                "enable_pickup": {
                    "type": "boolean"
                },
                "enable_reservations": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "description": "Configuración de marca blanca",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary_color": {
                    "type": "string"
                },
                "secondary_color": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.BusinessSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.BusinessResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingrese su token JWT con el prefijo **Bearer**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}